requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
EZSheets==2023.8.11
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
jdcal==1.4.1
MouseInfo==0.1.3
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.82
openpyxl==2.6.2
outcome==1.3.0.post0
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.3
pyobjc-framework-Cocoa==10.3
pyobjc-framework-Quartz==10.3
pyparsing==3.1.2
PyPDF2==1.26.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytweening==1.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
rubicon-objc==0.4.9
selenium==4.19.0
Send2Trash==1.8.2
setuptools==70.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==1.26.15
wheel==0.43.0
wsproto==1.2.0
